Add Segment::offset method
Add a method for segments which returns the on-disk file offset and
segment size (distinct from the in-memory size, for formats where they
can differ).

This is similar to `data()`, but rather than returning a slice of the
data itself, it returns the location of the data.  This is useful for
when we want to locate the segment within a memory mapping, or use it
to generate the parameters for an IO syscall.