Use SymbolSection::None for file symbols
For ELF, file symbols were being emitted with STB_GLOBAL instead
of STB_LOCAL. This was because their section was set to `Undefined`,
which overrode the symbol scope.

Fix by using a new `SymbolSection::None` in `add_file_symbol`.

This also changes the section to None when reading, which makes our
round trips consistent.