Fix compression feature
The behaviour of methods should not depend on a cargo feature.
So we give `ObjectSection::uncompressed_data` fixed behaviour,
but make its presence depend on the feature.

This requires making the traits sealed, which we probably should
do anyway.

Also make it possible to handle decompression errors.
If the user wants to ignore errors then they should do that
themselves.