Switch to wasmparser
This switches reader from manual implementation to wasmparser in order to prepare for implementation of new features and better integration like #196.

For the simplicity of review, this changeset intentionally doesn't implement any new features yet nor changes public API, and instead preserves same functionality but with wasmparser instead of a manual reader.