Make `target-lexicon` an optional dependency
This commit should finish out #215 by removing the dependencies of the
`object` crate if it's built with few enough features. While hopefully
not exercised by most users this should be enough for the `backtrace`
crate to depend on `object` and not pull in any transitive dependencies.
One step closer to moving into libstd!

Closes #215