Replace `Symbol` with a trait
This allows users to only store the parts of the symbol that they
are interested in. For example, the user may want to only store
the symbol index and retrieve specific details from the symbol table
only as needed.

`Object::symbol_map` now returns a map containing only symbol names and
addresses, which is the most common use case. `SymbolMap` can be
constructed with custom symbol entries if other information is required.
This method also uses the dynamic symbol table if there are no
debugging symbols.

Also fixes COFF/PE symbols to give the correct address. Previously
they weren't including the image base address and the section address.