Quickfix: Ignore errors when loading name subsection. (#403)
The version of `wasmparser` currently in use does not
support extended name sections [1] and will return an error
when such an unknown subsection is encountered.
This commit is a quick work-around until `wasmparser`
is updated to the most recent version - which is a
rather big undertaking unfortunately.

One downside of this work-around is that ill-formed
name subsections (e.g. for function names) are
silently ignored.

[1] (https://github.com/WebAssembly/extended-name-section/blob/main/proposals/extended-name-section/Overview.md)