read/wasm: Ignore errors when loading name subsection (#408)
The version of `wasmparser` currently in use does not
support extended name sections and will return an error
when such an unknown subsection is encountered.
This commit is a quick work-around until `wasmparser`
is updated to the most recent version - which is a
rather big undertaking unfortunately.

One downside of this work-around is that ill-formed
name subsections (e.g. for function names) are
silently ignored.