elf: Update LoongArch eflags (#483)
In a revision of LoongArch ELF ABI [1], it was decided that ILP32
objects must use ELFCLASS32.  Now both LP64D and ILP32D will use the
same number (0x1) for "double float", so EF_LARCH_ABI_LP64{S,F,D} are
renamed and EF_LARCH_ABI_ILP32{S,F,D} are removed.

The change is ABI-compatible for LP64, but not API-compatible.  ILP32 is
not supported by GCC and Glibc yet so it's not really used anyway.

And, a separate bit is allocated to indicate if the ELF object file uses
any relocation type directly writing to immediate slots [2].

Update the code to match these changes.

[1]:https://github.com/loongson/LoongArch-Documentation/pull/47/files#diff-711b3e7b6a005b492898ac6d93f2d8d37c00e0831e210993a6f9dbb26c043717
[2]:https://github.com/loongson/LoongArch-Documentation/pull/61/files#diff-711b3e7b6a005b492898ac6d93f2d8d37c00e0831e210993a6f9dbb26c043717