elfcopy: Add --redefine-syms <file> arguments (#508)
This commit makes elfcopy the only tool I'm aware of that will let you
rename symbols in dynamic libraries. Other tools like objcopy will only
perform renaming on the symbol table, but will not rename dynamic
strings or recalculate the hashes/offsets that change as a result.

This works by creating a redefinition table, and handing symbol (dynamic
and otherwise) to the table's get_redefined_name function before adding
that string to the out_* variables. The elf binary is then regenerated
from the output data (containing the renamed strings) and a valid DSO is
generated.