macho: fix PC relative relocation offset for AArch64
For PC relative relocations on some architectures, the implicit addend
does not include the offset required due to the PC being different from
the place of the relocation.

Previously we always used an offset of 4, since this is commonly required
by x86-64. However, this is not correct for AArch64 or for some x8-64
relocations, and probably not correct for other architectures either.
For example, the ARM offset may vary depending on the instruction used.

For now, disable this handling for all non x86 targets. This fixes AArch64.
Support for other architectures is left as a TODO.