read/coff: improve detection of auxiliary symbols for sections
Previously we wrongly excluded section symbols with non-zero values,
and wrongly included function symbols with zero values.

This also means that we give them correct symbol kind,
and that we exclude them from the symbol map.

Note that we still don't give the correct symbol kind to
section symbols that don't have an auxiliary symbol.