Allow generation of all output testfiles
Some of the output testfiles are only the partial output.
Extend the OBJECT_TESTFILES_UPDATE support to handle these too.

Also, reverse the derivation of filenames: now the input filename
is derived from the output filename. This allows us to detect
output files that don't have a corresponding input file.