Implement `core::error::Error` on Rust 1.81+ (#747)
This commit updates when the `Error` trait is implemented to account for
how in Rust 1.81-and-prior the `Error` trait is always available, even
in `core`. This is currently done with detection at build-time of the
current Rust compiler version because the MSRV for this crate is below
1.81. In the future when the MSRV is increased, however, the build
script can be deleted.