read/coff: fix has_aux_function for weak externals (#772)
has_aux_function previously matched weak externals too, resulting in
a wrong interpretation of their auxiliary symbol in the readobj
example.

Removed use of `else` for checking auxiliary symbols in readobj
so that it is more apparent when these are wrong.

Also add objdump and readobj test output for weak externals.