Add helper methods for mutating section and symbol flags
Doing it the naive way with first getting a mutable reference to the
section/symbol, then writing the result of section_flags/symbol_flags to
replace SectionFlags::None and SymbolFlags::None and then mutating the
flags as needed isn't allowed by the borrow checker. Instead you have to
do a dance with getting the section/symbol as immutable, then calling
section_flags/symbol_flags and then getting it as mutable and writing
the flags you got. Adding helper methods to object saves users the time
to figure this all out for themselves.